+++ processor/src/main/kotlin/com/infinum/collar/processor/extensions/Element.kt
@@ -1,10 +1,9 @@
 package com.infinum.collar.processor.extensions
 
 import com.squareup.kotlinpoet.ClassName
-import com.squareup.kotlinpoet.metadata.KotlinPoetMetadataPreview
 import com.squareup.kotlinpoet.metadata.isClass
 import com.squareup.kotlinpoet.metadata.isSealed
-import com.squareup.kotlinpoet.metadata.toImmutableKmClass
+import com.squareup.kotlinpoet.metadata.toKmClass
 import javax.lang.model.element.Element
 import javax.lang.model.element.ElementKind
 import javax.lang.model.element.NestingKind
@@ -13,18 +12,16 @@ import javax.lang.model.element.TypeElement
 import javax.lang.model.element.VariableElement
 import javax.lang.model.util.ElementFilter
 
-@KotlinPoetMetadataPreview
 internal fun Element.isSealedClass(): Boolean =
     this.getAnnotation(Metadata::class.java)
-        .toImmutableKmClass()
+        .toKmClass()
         .run {
-            this.isClass && this.isSealed
+            this.isClass && this.flags.isSealed
         }
 
-@KotlinPoetMetadataPreview
 internal fun Element.constructorParameterNames(): List<String> =
     this.getAnnotation(Metadata::class.java)
-        .toImmutableKmClass()
+        .toKmClass()
         .constructors
         .firstOrNull()
         ?.valueParameters