+++ processor/src/main/kotlin/com/infinum/collar/processor/validators/AnalyticsEventsValidator.kt
@@ -1,22 +1,21 @@
 package com.infinum.collar.processor.validators
 
+import com.infinum.collar.processor.extensions.allReservedKeywords
 import com.infinum.collar.processor.extensions.isSealedClass
 import com.infinum.collar.processor.extensions.showError
 import com.infinum.collar.processor.extensions.showWarning
 import com.infinum.collar.processor.models.AnalyticsEventsHolder
 import com.infinum.collar.processor.models.EventHolder
 import com.infinum.collar.processor.options.Options
-import com.squareup.kotlinpoet.metadata.KotlinPoetMetadataPreview
 import javax.annotation.processing.Messager
 import javax.lang.model.util.Types
 
 internal class AnalyticsEventsValidator(
-    private val processorOptions: Options,
+    private val processorOptions: Options?,
     private val typeUtils: Types,
     private val messager: Messager
 ) : Validator<AnalyticsEventsHolder> {
 
-    @KotlinPoetMetadataPreview
     override fun validate(elements: Set<AnalyticsEventsHolder>): Set<AnalyticsEventsHolder> {
         val validRootClasses = elements.filter {
             if (it.rootClass.isSealedClass().not()) {
@@ -27,7 +26,7 @@ internal class AnalyticsEventsValidator(
             }
         }
         val innerClassesCount = validRootClasses.flatMap { it.eventHolders }.count()
-        return if (innerClassesCount > processorOptions.maxCount()) {
+        return if (processorOptions != null && innerClassesCount > processorOptions.maxCount()) {
             messager.showError(
                 "You can report up to ${processorOptions.maxCount()} different events per app. " +
                     "Current size is $innerClassesCount."
@@ -53,7 +52,7 @@ internal class AnalyticsEventsValidator(
         }
 
     private fun validateNameLength(holder: EventHolder): Boolean =
-        if (holder.eventName.length > processorOptions.maxNameSize()) {
+        if (processorOptions != null && holder.eventName.length > processorOptions.maxNameSize()) {
             messager.showWarning(
                 "Event names can be up to ${processorOptions.maxNameSize()} characters long. " +
                     "${holder.eventName} is ${holder.eventName.length} long."
@@ -87,6 +86,7 @@ internal class AnalyticsEventsValidator(
 
     private fun validateReserved(holder: EventHolder): Boolean =
         if (
+            processorOptions != null &&
             processorOptions
                 .reservedPrefixes()
                 .any { holder.eventName.startsWith(it, false) }.not() && processorOptions.reserved()
@@ -95,19 +95,14 @@ internal class AnalyticsEventsValidator(
             validateParameterCount(holder)
         } else {
             messager.showWarning(
-                "The ${
-                processorOptions
-                    .reservedPrefixes()
-                    .plus(processorOptions.reserved())
-                    .joinToString { "\"$it\"" }
-                }" +
+                "The ${processorOptions?.allReservedKeywords().orEmpty()}" +
                     " event names are reserved and cannot be used as in ${holder.eventName}."
             )
             false
         }
 
     private fun validateParameterCount(holder: EventHolder): Boolean =
-        if (holder.eventParameters.size > processorOptions.maxParametersCount()) {
+        if (processorOptions != null && holder.eventParameters.size > processorOptions.maxParametersCount()) {
             messager.showWarning(
                 "You can associate up to ${processorOptions.maxParametersCount()} unique parameter with each event. " +
                     "Current size is ${holder.eventParameters.size}."