+++ processor/src/main/kotlin/com/infinum/collar/processor/validators/UserPropertiesValidator.kt
@@ -1,22 +1,21 @@
 package com.infinum.collar.processor.validators
 
+import com.infinum.collar.processor.extensions.allReservedKeywords
 import com.infinum.collar.processor.extensions.isSealedClass
 import com.infinum.collar.processor.extensions.showError
 import com.infinum.collar.processor.extensions.showWarning
 import com.infinum.collar.processor.models.PropertyHolder
 import com.infinum.collar.processor.models.UserPropertiesHolder
 import com.infinum.collar.processor.options.Options
-import com.squareup.kotlinpoet.metadata.KotlinPoetMetadataPreview
 import javax.annotation.processing.Messager
 import javax.lang.model.util.Types
 
 internal class UserPropertiesValidator(
-    private val processorOptions: Options,
+    private val processorOptions: Options?,
     private val typeUtils: Types,
     private val messager: Messager
 ) : Validator<UserPropertiesHolder> {
 
-    @KotlinPoetMetadataPreview
     override fun validate(elements: Set<UserPropertiesHolder>): Set<UserPropertiesHolder> {
         val validRootClasses = elements.filter {
             if (it.rootClass.isSealedClass().not()) {
@@ -27,7 +26,7 @@ internal class UserPropertiesValidator(
             }
         }
         val innerClassesCount = validRootClasses.flatMap { it.propertyHolders }.count()
-        return if (innerClassesCount > processorOptions.maxCount()) {
+        return if (processorOptions != null && innerClassesCount > processorOptions.maxCount()) {
             messager.showError(
                 "You can report up to ${processorOptions.maxCount()} different user properties per app. " +
                     "Current size is $innerClassesCount."
@@ -53,7 +52,7 @@ internal class UserPropertiesValidator(
         }
 
     private fun validateNameLength(holder: PropertyHolder): Boolean =
-        if (holder.propertyName.length > processorOptions.maxNameSize()) {
+        if (processorOptions != null && holder.propertyName.length > processorOptions.maxNameSize()) {
             messager.showWarning(
                 "User property name can be up to ${processorOptions.maxNameSize()} characters long. " +
                     "${holder.propertyName} is ${holder.propertyName.length} long."
@@ -87,6 +86,7 @@ internal class UserPropertiesValidator(
 
     private fun validateReserved(holder: PropertyHolder): Boolean =
         if (
+            processorOptions != null &&
             processorOptions
                 .reservedPrefixes()
                 .any { holder.propertyName.startsWith(it, false) }.not() && processorOptions.reserved()
@@ -95,12 +95,7 @@ internal class UserPropertiesValidator(
             validateExistingParameters(holder)
         } else {
             messager.showWarning(
-                "The ${
-                processorOptions
-                    .reservedPrefixes()
-                    .plus(processorOptions.reserved())
-                    .joinToString { "\"$it\"" }
-                }" +
+                "The ${processorOptions?.allReservedKeywords().orEmpty()}" +
                     " user properties are reserved and cannot be used as in ${holder.propertyName}."
             )
             false
@@ -110,15 +105,19 @@ internal class UserPropertiesValidator(
         if (holder.propertyParameterNames.isNotEmpty()) {
             validateParameterCount(holder)
         } else {
-            messager.showWarning(
-                "You must associate at least ${processorOptions.maxParametersCount()} unique parameter. " +
-                    "${holder.propertyName} has none."
-            )
-            false
+            if (processorOptions != null) {
+                messager.showWarning(
+                    "You must associate at least ${processorOptions.maxParametersCount()} unique parameter. " +
+                        "${holder.propertyName} has none."
+                )
+                false
+            } else {
+                true
+            }
         }
 
     private fun validateParameterCount(holder: PropertyHolder): Boolean =
-        if (holder.propertyParameterNames.size > processorOptions.maxParametersCount()) {
+        if (processorOptions != null && holder.propertyParameterNames.size > processorOptions.maxParametersCount()) {
             messager.showWarning(
                 "You can associate up to ${processorOptions.maxParametersCount()} " +
                     "unique parameter with each user property. " +